チャットPanelを表示componentへ分離する
